# docs/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/auth/refresh":{"get":{"operationId":"AuthController_refreshToken","summary":"","description":"","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenResponse"}}}}},"tags":["Auth"]}},"/api/auth/check":{"post":{"operationId":"AuthController_checkToken","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckTokenRequest"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/user/sign-in":{"post":{"operationId":"UserController_signIn","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInRequest"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponse"}}}}},"tags":["Users"]}},"/api/user/sign-up":{"post":{"operationId":"UserController_signUp","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequest"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpResponse"}}}}},"tags":["Users"]}},"/api/user/logout":{"post":{"operationId":"UserController_logout","summary":"","description":"","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutResponse"}}}}},"tags":["Users"]}},"/api/user/account-confirmation/{confirmHash}":{"get":{"operationId":"UserController_accountConfirmation","summary":"","description":"","parameters":[{"name":"confirmHash","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/user/get-settings":{"get":{"operationId":"UserController_getSettings","summary":"","description":"","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetSettingsResponse"}}}}},"tags":["Users"]}},"/api/user/freeze-account":{"patch":{"operationId":"UserController_freezeAccount","summary":"","description":"","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FreezeAccountResponse"}}}}},"tags":["Users"]}},"/api/user/close-account":{"patch":{"operationId":"UserController_closeAccount","summary":"","description":"","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloseAccountResponse"}}}}},"tags":["Users"]}},"/api/phone/unbind":{"post":{"operationId":"PhoneController_unbindPhone","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnbindPhoneRequest"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnbindPhoneResponse"}}}}},"tags":["Phone"]}},"/api/phone/confirm":{"post":{"operationId":"PhoneController_confirmPhone","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhoneConfirmationRequest"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhoneConfirmationResponse"}}}}},"tags":["Phone"]}},"/api/phone/verify":{"post":{"operationId":"PhoneController_verifyPhone","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyPhoneRequest"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyPhoneResponse"}}}}},"tags":["Phone"]}},"/api/twofactor/set":{"post":{"operationId":"TwofactorController_setTwoFa","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetTwoFaRequest"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetTwoFaResponse"}}}}},"tags":["2FA"]}},"/api/twofactor/remove":{"post":{"operationId":"TwofactorController_removeTwoFa","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemoveTwoFaRequest"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemoveTwoFaResponse"}}}}},"tags":["2FA"]}}},"info":{"title":"Cryptodiscrict API","description":"Cryptodiscrict API documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"SessionDto":{"type":"object","properties":{"id":{"type":"string"},"tokenId":{"type":"string"},"userId":{"type":"string"},"updatedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"}},"required":["id","tokenId","userId","updatedAt","createdAt"]},"RefreshTokenResponse":{"type":"object","properties":{"_at":{"type":"string"}},"required":["_at"]},"CheckTokenRequest":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"SignInRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","minLength":6,"maxLength":32}},"required":["email","password"]},"SignInResponse":{"type":"object","properties":{"_at":{"type":"string"}},"required":["_at"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"twitter":{"type":"string"},"linkedIn":{"type":"string"},"personalWebsite":{"type":"string"},"title":{"type":"string"},"bio":{"type":"string"},"tac":{"type":"boolean"},"accountConfirm":{"type":"boolean"},"updatedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"emailChanged":{"type":"boolean"},"lastPassChange":{"format":"date-time","type":"string"},"twoFaType":{"type":"object"},"receiveNotifications":{"type":"boolean"}},"required":["id","firstName","lastName","phoneNumber","email","password","twitter","linkedIn","personalWebsite","title","bio","tac","accountConfirm","updatedAt","createdAt","emailChanged","lastPassChange","twoFaType","receiveNotifications"]},"SignUpRequest":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string","minLength":8},"twitter":{"type":"string"},"linkedIn":{"type":"string"},"personalWebsite":{"type":"string"},"title":{"type":"string"},"bio":{"type":"string"},"tac":{"type":"boolean"}},"required":["email","password","tac"]},"SignUpResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"LogoutResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"ConfirmHashDto":{"type":"object","properties":{"id":{"type":"string"},"confirmHash":{"type":"string"},"userId":{"type":"string"},"confirmed":{"type":"boolean"},"updatedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"}},"required":["id","confirmHash","userId","confirmed","updatedAt","createdAt"]},"GetSettingsResponse":{"type":"object","properties":{"settings":{"type":"object"}},"required":["settings"]},"FreezeAccountResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"CloseAccountResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"UnbindPhoneRequest":{"type":"object","properties":{}},"UnbindPhoneResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"PhoneDto":{"type":"object","properties":{"id":{"type":"string"},"phone":{"type":"string"},"userId":{"type":"string"},"updatedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"}},"required":["id","phone","userId","updatedAt","createdAt"]},"VerificationCodeDto":{"type":"object","properties":{"id":{"type":"string"},"verificationCode":{"type":"string"},"userId":{"type":"string"},"confirmed":{"type":"boolean"},"updatedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"}},"required":["id","verificationCode","userId","confirmed","updatedAt","createdAt"]},"PhoneConfirmationRequest":{"type":"object","properties":{"code":{"type":"string","minLength":6,"maxLength":6}},"required":["code"]},"PhoneConfirmationResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"VerifyPhoneRequest":{"type":"object","properties":{}},"VerifyPhoneResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"TwoFaDto":{"type":"object","properties":{"id":{"type":"string"},"twoFaToken":{"type":"string"},"userId":{"type":"string"},"updatedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"}},"required":["id","twoFaToken","userId","updatedAt","createdAt"]},"SetTwoFaRequest":{"type":"object","properties":{}},"SetTwoFaResponse":{"type":"object","properties":{}},"RemoveTwoFaRequest":{"type":"object","properties":{}},"RemoveTwoFaResponse":{"type":"object","properties":{}}}}}